requirements.txt
qwen-vl-utils>=0.0.8
torchvision
transformers>=4.47.0
accelerate>=1.1.1
huggingface_hub
spaces
gradio>=5.6.0
deepspeed
wandb
peft
tensorboard
tqdm
opencv-python
ipython
pillow
